Move datastore helper functions to shared file
